requirements.txt
agate==1.9.1
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
Authlib==1.6.8
avro==1.12.1
babel==2.18.0
backrefs==6.2
cachetools==7.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
clickhouse-connect==0.11.0
clickhouse-driver==0.2.10
colorama==0.4.6
confluent-kafka==2.13.0
cryptography==46.0.5
daff==1.4.2
dbt-adapters==1.22.5
dbt-clickhouse==1.10.0
dbt-common==1.37.2
dbt-core==1.11.6
dbt-extractor==0.6.0
dbt-protos==1.0.431
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.1
fastavro==1.12.1
ghp-import==2.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kafka-python==2.3.0
leather==0.4.1
lz4==4.4.5
Markdown==3.10.2
MarkupSafe==3.0.3
mashumaro==3.14
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.7.2
mkdocs-material-extensions==1.3.1
more-itertools==10.8.0
msgpack==1.1.2
networkx==3.6.1
numpy==2.4.2
orderly-set==5.5.0
orjson==3.11.7
packaging==26.0
paginate==0.5.7
pandas==3.0.1
parsedatetime==2.6
pathspec==0.12.1
platformdirs==4.9.2
protobuf==6.33.5
py4j==0.10.9.9
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymdown-extensions==10.21
pyrsistent==0.20.0
pyspark==4.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
pyyaml_env_tag==1.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
snowplow-tracker==1.1.0
sqlparse==0.5.4
text-unidecode==1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
watchdog==6.0.0
zipp==3.23.0
zstandard==0.25.0
xgboost==2.1.3
scikit-learn==1.6.1
joblib==1.4.2
matplotlib
